Restructure test architecture to better fit the structure
